+++ src/eventcontroller/AudioVideoEventAttributes.ts
@@ -13,4 +13,5 @@ export default interface AudioVideoEventAttributes {
   retryCount?: number;
   signalingOpenDurationMs?: number;
   iceGatheringDurationMs?: number;
+  selfAttendeePresentDurationMs?: number;
 }